* Declaration.java (flags): Chaneg from int to long. 	(getFlag, setFlag, printInfo): Update accordingly. 	* LambdaExp.java (compileAsMethod): Update accordingly. 	(VOLATILE_ACCESS, TRANSIENT_ACCESS): New flags. 	(getAccessFlags): Re-write to support new flags.
	* standard/object.java (matchAccess): Re-write to support
	volatile and transient flags, and also a list of multiple flags.
	(scanClassDef): Update accordingly.
	* lang/Translator.java (matchQuoted): New helper method.


git-svn-id: svn://sourceware.org/svn/kawa/trunk@6444 5e0a886f-7f45-49c5-bc19-40643649e37f

+++ gnu/expr/LambdaExp.java
@@ -1504,10 +1504,10 @@ public class LambdaExp extends ScopeExp
     int numStubs = primMethods.length - 1;
     Type restArgType = restArgType();
 
-    int[] saveDeclFlags = null;
+    long[] saveDeclFlags = null;
     if (numStubs > 0)
       {
-	saveDeclFlags = new int[min_args + numStubs];
+	saveDeclFlags = new long[min_args + numStubs];
 	int k = 0;
 	for (Declaration decl = firstDecl();
 	     k < min_args + numStubs; decl = decl.nextDecl())
